+++ Items/Dev/RuneBook/VoidRune.cs
@@ -234,7 +234,7 @@ namespace AAMod.Items.Dev.RuneBook
 						Vector2 value19 = vector46 - projectile.Center;
 						value19.Normalize();
 						value19 *= scaleFactor3;
-						int num659 = Projectile.NewProjectile(projectile.Center.X, projectile.Center.Y, value19.X, value19.Y, num658, (int)(projectile.damage), 0f, Main.myPlayer, projectile.whoAmI, target);
+						int num659 = Projectile.NewProjectile(projectile.Center.X, projectile.Center.Y, value19.X, value19.Y, num658, projectile.damage, 0f, Main.myPlayer, projectile.whoAmI, target);
 						projectile.netUpdate = true;
 					}
 				}
